Aggregate's state must be internal and not accessible from outside.
The only internal attribute available to read from outside is an aggregate's id.
This is needed to know what aggregate should we load/store.

All others attrubites (state here in the example) must be protected and changed only
by applying an event. And because they are changed only by applying an event we do not
meet to test state values because all we need it to check if valid event have been produced.