Refresh docs/site design.
Initially I only wanted to tweak fonts on railseventstore.org. This
turned out to be far more complicated that it should - in gitbook any
customization like this is done via plugin/theme.

Gitbook, while cool enough for just docs, seems to be less suitable to
expose project landing page and describe how this project might be
useful to broader audience. We likely won't be generating books from
the docs as well.

I gave a quick look to https://github.com/PharkMillups/beautiful-docs
and noticed middlemanapp.com - both as a great showcase for project
site/docs and as a tool to achieve that goal.

This commit introduces:
- new site design with docs and landing pages
- generating site with middleman over gitbook
- minor copy rewording
- CD pipeline to build + deploy docs on source file change so that no
one is bothered to have toolchain locally to push new docs