New way of defining subscriptions
Extension on the subscriptions API. The "old" `subscribe(...)` method
stays unchagnged but 2 new methods are defined:

* `on(event_types, subscriber_klass, &proc)` - allow to add subscription
  class for given set of event types
* `on_all_events(subscriber_klass, &proc)` - allow to add subscription
  class for all event types

It builds a proxy that will call a method (#call on `DefaultDispatcher`)
on a `subscriber_klass` and this proxy will be used as argument to "old"
subscribe method.

The source of proxy is a dispatcher assigned to event broker. That will
allow to plug in different dispatchers and diffrent proxy strategies.