Extracted a private EventRepositoryReader class
I have collected some metrics/facts/opinions about the EventRepository
class and this is one possible attempt to address some existing issues.

- rich public API - 14 public methods
- 259 Lines of Code
- 1 class
- several responsibilities

responsibilities

- detecting if correct schema and display message
- dealing with the mapper
- raising the proper exceptions
- versioning
- detecting MySQL, Postgres, Sqlite3 index violations
- building event records and event instances
- normalising input
- appending to stream (implicitly adds to global stream too)
- linking to stream (doesn't add to global stream)
- deleting a stream
- has_event check
- readers:
	- last_event_stream
	- read_events_forward
	- read_events_backward
	- read_stream_events_forward
	- read_all_streams_forward
	- read_all_streams_backward
	- read_event
	- get_all_streams

The actual change is rather cosmetical here - all the reader methods
are extracted and delegated to another class, in the spirit of CQS.

It doesn't change any public API, it's just an attempt to make the
EventRepository logic easier (subjective) to reason about.

Possible next steps are to extract a EventRepositoryWriter class too.