Extracted a private CorrectSchemaVerifier class
This responsibility sounds like a good fit for a new class so that we
can simplify the EventRepository class.

I have assumed that InvalidDatabaseSchema is part of our public API,
so I kept it within the same EventRepository namespace.

I have extracted smaller methods to make it clearer how this algorithm
works in detail - compares the columns.