Fix Rails 4.2 duck typing ;)
In Rails 4.2 #rolledback! is called with 1-2 boolean attributes.

      def rollback_records
        ite = records.uniq
        while record = ite.shift
          begin
            record.rolledback! full_rollback?
          rescue => e
            raise if ActiveRecord::Base.raise_in_transactional_callbacks
            record.logger.error(e) if record.respond_to?(:logger) && record.logger
          end
        end
      ensure
        ite.each do |i|
          i.rolledback!(full_rollback?, false)
        end
      end

In rails 5 it is called with 1-2 named attributes

      def rollback_records
        ite = records.uniq
        while record = ite.shift
          record.rolledback!(force_restore_state: full_rollback?)
        end
      ensure
        ite.each do |i|
          i.rolledback!(force_restore_state: full_rollback?, should_run_callbacks: false)
        end

Taking dynamic list of arguments (*) instead of named arguments (**)
handles both cases.

ActiveRecord::Base.raise_in_transactional_callbacks was removed in
Rails5 (moving to behavior like setting it to true). So the test detects the
situation and only executes in 1 environment to proove that the
interface is correct. Otherwise Rails catches our exception and swallows
it so we don't see that we have an incorrect number of arguments.

Refs: #181