Encoded names on boundaries (URLs), decoded in core.
We made a mistake of decoding stream name just for the view, which made
it look good. When passing undecoded stream name to Cmd, we'd encode it
one again, making it unrecognizable to backend.

Now we decode names after URL parsing and encode whenever URL is needed
(http request or anchor tag).