reverse() is not necessary
because the generated method name (value in Hash) in inherited class is the same.
So it does not matter if we merge the hash from top to bottom or
otherwise.

It mattered in my initial idea that I had in which the value was the method
itself instead of the method's name. Mutation testing tells me so.

Refs: #171 #190