PoC - new and safe dynamic subscriptions API
Proof of concept

Before calling within current event_broker is deeply cloned and put on
top of thread-scoped stack. Within the #within() method subscription
manipulation operates on that copy. When the block finishes we bring
back the old object.

I believe (to be tested) this should be safe in case multiple threads
(puma etc) doing multiple #within operations simulataniously.

Assumption: It's ok that other threads manipulations to subscribers are
not visible when the thread is inside #within block.

Thread 1: [Broker A]
Thread 2: [Broker A, Broker A+B]
Thread 3: [Broker A, Broker A+C, Broker A+C+D]

Initially I wanted to add

* add_temporary_subscriber
* add_temporary_global_subscriber
(maybe thread_local instead of temporary would be a better name)
and then maybe refactor to a single method with some modifiers (global
or limited to some events, thread_local or not).

to Broker and make Broker handle those subscriptions in
Concurrent::ThreadLocalVar but I realized that maybe keeping it
thread-unware could be a better solution. Now it only knows how to
duplicate itself so that the copy can be enriched with temporary
subscriptions.

Is this cool or crazy?

Refs: #196 #155