Test it does not not expose uncommited events
Nothing needs to be changed for it to work as default PostgreSQL
transaction isolation level is READ COMMITTED. It does not even support
READ UNCOMMITTED so this just works out of the box.

But we now have a test setup that we can use for testing the real
troubling cases that we want to support.