Messy commit
Scenario that I expected to fail passed because we use
ActiveRecord::Base.transaction(requires_new: true) inside the repository
which works like a new transaction and bumps xmin.

This is not true in the "debugging" cases where I manually insert record
in the same transaction.

That does not mean the code is ready. I means we need to figure out a
better test case scenario.