Handle possible interleaved row IDs.
When debugging the most challenging test:

   # Thread1: [1] [2,      4]
   # Thread2:          [3,     ]

I was surprised to learn that it passes without any code change.
Upon investigating it turned that this was caused by

  ActiveRecord::Base.transaction(requires_new: true)

which creates SAVEPOINT which bumps xmin of inserted event recrods.

I wanted to test a possible situation when there are 2 transactions
inserting at the same time multiple events and they get interleaving IDs
and mixed xmin. This was hard to achieve coventionally even with
thousands of records because we use EventInStream.import(in_stream)
in our repository which at the end inserts events to a stream with
a single statement that takes miliseconds and it is hard to synchronize
it between two threads (idea: maybe we need 20 threads?).

Also: I am not sure if that is even possible frankly. If not, we could
probably just throw away this code. But I really could not find more
info about it. What's IT? Is it possible that when 2 transactions
concurrently insert multiple records in a single statement that they
would get interleaved IDs from autoincrement? If it is not possible,
maybe we could just rely on the previous version of code. I don't know.

So here is what I did. I manually changed
ActiveRecord::Base.transaction(requires_new: true)
to
ActiveRecord::Base.transaction(requires_new: false)
and run the tests. That lead me to current solution.

TODO: Make those 2 versions in tests.

Now let's talk about the current solution.

We want the client to be able to safely, incrementally iterate over
events in global ('all') stream. That means when we get events after
particular previous event we need to make sure that there is no GAP in
the returned slots. The GAP can occur when there is an ongoing
uncommitted transaction which could have created an event with certain
number. The GAP can also occur when there was a rollback. Or when the ID
was taken for placing the event in a different stream.

The situation of ongoing transaction and different stream can occur
simulataniusly and we cannot distinguish it because we don't have a way
to peak into other transactions in a simple way.

Here is what we do in every situation:

* if ID is already found and just next one in a sequence, that means the
event can be synced.
* if not, we check if maybe it is simply taken in non-global stream and
we can skip it.
* if not (last resort, that should be super rare) we try to detect if it
was rolledback or is used in an ongoing transaction. If rolled-back then
we can move on. If not, we need to stop and delay further
synchronization until the transaction is finished.

How do we recognize the GAP is from a rolledback transaction? The
sequence number can be taken in an INSERT and it will pass without a
problem.

If the transaction JUST commited then ON CONFLICT (id) DO NOTHING will
detect a conflict on primary key during INSERTion and we will get 0
as number of inserted records.

If the transaction is ongoing then unfortunately ON CONFLICT is not
helpful and Postgres will wait for transactions to finish. We want to
fail fast ("set statement_timeout to 100") in 100ms and detect that
situation. That means the sync ends here as we reached maximal
synchronization point until further transactions are committed.

But I believe that due to the before-mentioned SNAPSHOT behavior and the fact
that all events inserted in a single publish_event call seem to have
monotonically increasing IDs this should be pretty rare code to reach
and it should not affect performance much in practice.