Emulate READ COMMITED with advisory locks
Locks are handled between multiple open transactions/connections.

Writer obtains lock with the number of 1st ID it will insert into
sequence.

Reader can check by asking for lock if there is an ongoing transaction
inserting at between detected GAPs in sequence of EventInStream
incremental IDs.

In DistributedRepository IDs (and another lock) are obtained in short
global lock (around 0.0005s). That another lock is what makes asking by
Reader for an ongoing transaction possible. Obtaining these sequence
numbers in a global lock is what makes it possible to assume that when
there is a GAP in IDs, then the first number from running transaction is
used for the temporary transactional lock and can be detected by an
another outside transaction.