Workaround YAML serialization issues with protobuf
by handling it on Proto wrapper level.

Seriously approaching this would require having a :mapper exposed
in :dispatcher which would require having it exposed in :broker but it
is already used in :repository.

So either we move the dependency higher and pass it down through many
layers, or make it a responsibility of :repository which can delegate to
:mapper if it wants to but then we would need to pass :repository around
to :dispatcher so it could call #event_to_serialized_record when enquing
a job and #serialized_record_to_event when dequeing.

I am not yet convinced about it.

Two additional thoughts:

* maybe repository should return RubyEventStore::SerializedRecord
  and then the migher layer (Client) would use the :mapper to transform
  it to the right event object.

* similarly when saving we could do the same, the other way, Client
  would already use mapper to serialize the event and repository would
  receive RubyEventStore::SerializedRecord.

* The benefit would be that we could get RAW data from repository, not
  transform it to domain events and stream over network.

  Current solution require getting domain events (which are
  deserialized) and then serializing them again to send over wire.

Issue: #228 #255