Test with sidekiq
to be double sure that our serialization for async handlers works
properly.

Issue: #346

----- ActiveJob::Base.queue_adapter global -----

Unfortunately in rails 4.2 Active job queue adapter was a class variable
shared between all classes:

module ActiveJob
  module QueueAdapter #:nodoc:
    extend ActiveSupport::Concern
    module ClassMethods
      mattr_reader(:queue_adapter) { ActiveJob::QueueAdapters::InlineAdapter }
      def queue_adapter=(name_or_adapter)
        @@queue_adapter = \
          case name_or_adapter
          when :test
            ActiveJob::QueueAdapters::TestAdapter.new
          when Symbol, String
            load_adapter(name_or_adapter)
          else
            name_or_adapter if name_or_adapter.respond_to?(:enqueue)
          end
      end

as a result writing:

class MySidekiqHandler < ActiveJob::Base
  self.queue_adapter = :sidekiq

affected all tests :( and set the queue adapter to sidekiq for all
handlers and not this one particular only.

Rails 5.2 uses class_attribute from ActiveSupport:
* Declare a class-level attribute whose value is inheritable by subclasses.
* Subclasses can change their own value and it will not impact parent class.

but we need our tests to work on both versions right now.

So I decided to operate on ActiveJob::Base.queue_adapter in all test to
unify the behavior between all versions that we support.

----- Sidekiq testing -----

I did not plan originally to use Sidekiq::Testing as I wanted be as close
to real scenario as possible. But I could not find an easy way to run a
real worker that would get the jobs from redis and process them and then
quit.

This seems to be good enough as we go through the serialization process:

  class Client
    alias_method :raw_push_real, :raw_push

    def raw_push(payloads)
      if Sidekiq::Testing.fake?
        payloads.each do |job|
          job = Sidekiq.load_json(Sidekiq.dump_json(job))
          job.merge!('enqueued_at' => Time.now.to_f) unless job['at']
          Queues.push(job['queue'], job['class'], job)
        end
        true
      elsif Sidekiq::Testing.inline?
        payloads.each do |job|
          klass = Sidekiq::Testing.constantize(job['class'])
          job['id'] ||= SecureRandom.hex(12)
          job_hash = Sidekiq.load_json(Sidekiq.dump_json(job))
          klass.process_job(job_hash)
        end
        true
      else
        raw_push_real(payloads)
      end
    end
  end

I run Sidekiq::Worker.drain_all in a new thread to decouple from
thread-level global variables.

https://github.com/mperham/sidekiq/wiki/Testing

----- STDOUT -----

Sidekiq is unhappy about Rails.env when we require it:

if defined?(::Rails) && Rails.respond_to?(:env) && !Rails.env.test?
  puts("⛔️ WARNING: Sidekiq testing API enabled...")

I reassign $stdout to mute this warning which is not of value for us.