PgLinearizedEventRepository
repository which uses lock (released at the end of a transaction) to
guarantee only one batch of events written concurrently.

* works only with postgres
  * mysql advisory locks are not easily releasible at the end of a
    transaction

* global/all stream still does not have "position" ordered but we are
  planning to drop it anyway.

* not mutated because mutations would fail on non-pg connections

* reading query result just in case with `each{}`. Perhaps not
  necessary.

Workaround for #106 which should be suitable on greenfield projects
which only use short transactions.

Not recommended to use when you publish from within long transactions as
it can block other concurrent operations.

Throughput and performance not benchmarked yet.