Additional check for Integer expected_version.
[#477]

When passing an integer, check if it doesn't exceed last known stream
position. Otherwise you'd create a gap in stream positions and allow
other writer to fill it in, possibly messing the order of events.

The drawback is an additional query on write for that case.

That query is a subject to race condition — we don't serialize last
stream position read with correspoding append. That however should not matter
for the "expected_version > last stream position" check - the writer
typically relies on last know stream position from its last read
(see AggregateRoot#load).