Redefine subscriptions - breaking changes
Introducing 2 new classes, Subscription & GlobalSubscription.
When new subscription is added an instance of one of tis classes will
be a result of add method instead of unsubscribe proc - this is breaking
change.

Unsubscribe now could be handled by calling on returned subscription
object an unsubscribe method.

Also dispatcher is affected here. Instead of subscriber (object or class)
it will receive as a first argument an subscription object. If
subscriber is a class a subscription object is now responsible for
instantialize new subscriber object (that have been responsibility of a
dispatcher).