That's reality, of_type works this way by design
We do not even have an instance of event store client
(not event mentioning remapping transformation) when we
define handlers using `when` method. Then `read_scope`
only fetches events listed in handlers defined for the
projection.

We could change `handled_events` method and try to fetch all
old types from remapping transformation but it is not the correct way.