Use the public API, not the private one
i.e. super_diff does not have these diff_as_object/diff_as_string methods, but works
correctly with a diff one.

The ` + "\n"` is there to make default rspec differ to actually print
the diff, because he skips displaying it if the strings are single
line strings. Not a perfect thing, but if we would like super_diff to
work with it, then it's a necessary one.