CDN browser assets when gem used via bundler with git source
When you use ruby_event_store-browser gem from git source, it lacks CSS
and JS that built just before packaging gem.

We're not versioning output files, only the source.

In this commit we're enabling reaching for CDN hosted assets, that are
built for every commit, when local assets are not available. This makes
browser work, even when RES is used from git source.

Related:
https://github.com/RailsEventStore/rails_event_store/issues/583