AR#insert_all needs timestamps
Previously, with activerecord-import, timestamps were taken care of. We
do pass explicit created_at for event_store_events. It was decided to
leave AR-generated timestamp when adding entries to
event_store_events_in_streams.

Now, after migrating to native insert_all API, we must take care of this
timestamp ourselves.

Automatic timestamp insertion was added later — probably around Rails 7,
but we support earlier versions too and it is not a big deal to generate
this timestamp ourselves.

https://github.com/rails/rails/pull/43003

This change should not modify existing behaviour. In the future we might
want to cede timestamp creation on the database or decide to pass the
timestamp from Record/SerializedRecord.