Revert elm-css direction
Typed CSS is nice and also not needing to think about CSS output was
nice. However it's a dead end with Content-Security-Policy. Our aim with
Browser is to be as compatible with various apps as possible so it was
limiting Browser adoption too much.

https://github.com/rtfeldman/elm-css/issues/569#issuecomment-1162874307
https://github.com/rtfeldman/elm-css/pull/570#issuecomment-1162858656
https://arkency.slack.com/files/U025BA51D/F03LJ1UPZQV/audio_clip__2022-06-22_01_11_39_.m4a

Previous idea with nonce that did not work out:
https://arkency.slack.com/archives/C7B95EW3V/p1655828630714969

Related:
https: //github.com/RailsEventStore/rails_event_store/issues/1346

Co-authored-by: Szymon Fiedler <szymon.fiedler@gmail.com>