No idea how to achieve the same in Rails < 6.1
Rails 6.1 improved attribute API a lot. Especially, it allows passing a
Proc to be evaluated after the schema is loaded and allows determining
the original type based on db schema. This allows us to make a decision:
whether to serialized data on ActiveRecord level or skip it, hence we
already seralized it by the serializer given to EventRepository.

Co-authored-by: Łukasz Reszke <lukaszreszke93@gmail.com>