Using projection with event type defined by event's class
During an attempt to introduce event_type_resolver into projection I
wrote following test. The idea is that you can use two different
versions of event with class level method defining the event type.

In that case, and in usage presented as in test, the event_type_resolver
doesn't bring any value to the projection class. So for now I removed
it.

Perhaps it does make more sense if we're thinking of other way
of using projections?