Add PostGIS to the list of supported adapters
PostGIS is a spatial database extender for PostgreSQL database. To allow
its usage within application, it requires different adapter on to of pg
one — activerecord-postgis-adapter, which extends capabilities of the
former.

To distinguish between those two, different adapter name in database
configuration has to be provided (postgis instead of postgresql).
Our VerifyAdapter prevented use generators if adapter wasn't pg, mysql
or sqlite until now.

As long as postgis simply extends capabilities of postgresql, we don't
see an issue with allowing the use of it, especially when this check
exists only during use of generators for db migrations to set up RES
database structure properly.