Fix browser issue with Float::INFINITY in event data
Resolves: https://github.com/RailsEventStore/rails_event_store/issues/1829

The JSON specification (RFC 8259) does not support special
floating-point values like Infinity, -Infinity, or NaN.

It states that:

> Numeric values that cannot be represented as sequences of
> digits (such as Infinity and NaN) are not permitted.

Ruby's JSON.parse doesn't allow parsing values without explicitly
stating this:

    irb(main):003> JSON.parse('[Infinity]')
    /Users/fidel/.gem/ruby/3.3.6/gems/json-2.8.1/lib/json/common.rb:204:in `parse': unexpected token at 'Infinity]' (JSON::ParserError)
    	from /Users/fidel/.gem/ruby/3.3.6/gems/json-2.8.1/lib/json/common.rb:204:in `parse'
    	from (irb):3:in `<main>'
    	from <internal:kernel>:187:in `loop'
    	from /Users/fidel/.gem/ruby/3.3.6/gems/irb-1.14.1/exe/irb:9:in `<top (required)>'
    	from /Users/fidel/.gem/ruby/3.3.6/bin/irb:25:in `load'
    	from /Users/fidel/.gem/ruby/3.3.6/bin/irb:25:in `<main>'
    irb(main):004> JSON.parse('[Infinity]', allow_nan: true)
    => [Infinity]

However, when producing a JSON output, it behaves a bit differently:

    irb(main):002> JSON.dump(Float::INFINITY)
    => "Infinity"
    irb(main):003> JSON.dump({ value: Float::INFINITY })
    => "{\"value\":Infinity}"

When infinite value is a part of the structure it's presented as
Infinity which represents Float::INFINITY instance. It becomes
"Infinity" string while being dumped to JSON directly.

While Infinity,-Infinity and NaN are not the part of JSON specification,
it has been decided that api endpoint providing data for browser will
presented those as "Infinity", "-Infinity" and "Nan".