Require OpenStruct only where necessary
Since Ruby 3.5 ostruct won't come as bundled gem, it's currently triggering
warning in projects using RailsEventStore as it's required in PreserveTypes:

  - https://github.com/RailsEventStore/rails_event_store/issues/1824
  - https://bugs.ruby-lang.org/issues/20309
  - https://github.com/ruby/json/pull/565

We don't rely on OpenStruct, but just support it being present in events'
data and metadata. RailsEventStore::JSONClient has been modified to
support OpenStruct only if it's present in project's bundle — we assume
that if someone made a decision to use this library, it was an informed
decision.

Still, it has to become development dependency.