Drop Ruby 3.1 from test matrix
As Ruby 3.1 will reach its EOL in a few weeks
(end or March 2025 https://endoflife.date/ruby)
we drop it from test matrix as it is incompatible
with Rails 8.

It could be still tested on Ruby 3.1 but it would
make our test matrix preparation much more complex,
too complex to remove it in a few weeks.