Failing test case to illustrate problem
This is a perfectly valid ActiveJob API to handle inline job execution.
It fails because normally it would expect to pass the raw job arguments.

We decorate the `perform` method, so that our prepended method silently
deserializes the arguments before passing them as an event object.

Possible options to resolve this:

* serialize with events with global id, providing custom resolver for
  global id — AJ responsible to fetch instance of an event and passing
  it to perform

* decorate AJ#initialize — tricky because of the
  serialization/deserialization built-in (see earlier point)

* override perform_now class and instance methods — at least the class
  method is a viable route, need to remember about supporting callbacks
  and error handlers

https://github.com/RailsEventStore/rails_event_store/issues/1786
https://github.com/RailsEventStore/rails_event_store/issues/1334