Drop bundler isolation
This was added purely to run `bundler/inline` within a spec executed
under a different bundle. It was supposed to clear any previous bundler
influence in the subprocess. Unfortunately due to the way we run bundler
on CI (config --deploymment 'true') this still doesn't work 100%
correctly — and cannot resolve gems confliciting with the bundled ones
(uri, securerandom, ...).

So let's not pretend it would work and get rid of it.